Moved TypeService
